Revert "Merge pull request #358 from atlasapi/jackson-ID-serilaization-issue"
This reverts commit fb7fe54c55e1b73ce715f3cda1b30e1d0a841f5a, reversing
changes made to afe391e2d53df1a65675da7f2f94e6f9bee61232.

+++ atlas-core/src/main/java/org/atlasapi/entity/Id.java
@@ -53,7 +53,7 @@ public final class Id implements Comparable<Id> {
     private long longValue;
 
     private Id(Long longValue) {
-        this.longValue = longValue.longValue();
+        this.longValue = longValue;
     }
 
     private Id(long longValue) {
